[VACGOM-120] refactor: User Service Query, Command 서비스로 분리
