+++ src/main/kotlin/kr/co/vacgom/api/user/application/UserCommandService.kt
@@ -1,16 +1,14 @@
 package kr.co.vacgom.api.user.application
 
-import kr.co.vacgom.api.baby.application.BabyService
+import kr.co.vacgom.api.baby.application.BabyCommandService
 import kr.co.vacgom.api.baby.domain.Baby
 import kr.co.vacgom.api.babymanager.application.BabyManagerService
 import kr.co.vacgom.api.babymanager.domain.BabyManager
 import kr.co.vacgom.api.global.exception.error.BusinessException
-import kr.co.vacgom.api.s3.S3Service
 import kr.co.vacgom.api.user.domain.User
 import kr.co.vacgom.api.user.domain.enums.UserRole
 import kr.co.vacgom.api.user.exception.UserError
 import kr.co.vacgom.api.user.presentation.dto.SignupDto
-import kr.co.vacgom.api.user.presentation.dto.UserDto
 import kr.co.vacgom.api.user.repository.UserRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -18,13 +16,12 @@ import java.util.*
 
 @Service
 @Transactional
-class UserService(
+class UserCommandService(
     private val userTokenService: UserTokenService,
     private val userRepository: UserRepository,
     private val authService: AuthService,
-    private val babyService: BabyService,
+    private val babyCommandService: BabyCommandService,
     private val babyManagerService: BabyManagerService,
-    private val s3Service: S3Service,
 ) {
     fun signup(request: SignupDto.Request): SignupDto.Response {
         val registerToken = userTokenService.resolveRegisterToken(request.registerToken)
@@ -45,9 +42,8 @@ class UserService(
             )
         }
 
-
         val savedUser = userRepository.save(newUser)
-        val savedBabies = babyService.saveAll(newBabies)
+        val savedBabies = babyCommandService.saveAll(newBabies)
         val managers = savedBabies.map { baby ->
             BabyManager(
                 user = savedUser,
@@ -73,23 +69,5 @@ class UserService(
 
         authService.unlinkUser(findUser)
         userRepository.deleteById(userId)
-    }
-
-    fun getUserDetail(userId: UUID): UserDto.Response.UserDetail {
-        return userRepository.findById(userId)?.let {
-            UserDto.Response.UserDetail(
-                id = it.id,
-                nickname = it.nickname,
-                socialId = it.socialId,
-                provider = it.provider,
-                role = it.role,
-                createdAt = it.createdAt,
-            )
-        } ?: throw BusinessException(UserError.USER_NOT_FOUND)
-    }
-
-    @Transactional(readOnly = true)
-    fun getUserById(userId: UUID): User {
-        return userRepository.findById(userId) ?: throw BusinessException(UserError.USER_NOT_FOUND)
     }
 }
